Simplify inline session error handling
