+++ packages/app/src/app/types.ts
@@ -67,13 +67,6 @@ export type MessageWithParts = {
   parts: Part[];
 };
 
-export type SessionErrorTurn = {
-  id: string;
-  text: string;
-  afterMessageID: string | null;
-  time: number;
-};
-
 export const SYNTHETIC_SESSION_ERROR_MESSAGE_PREFIX = "session-error:";
 
 export type StepGroupMode = "exploration" | "standalone";
